feat: rename address to address-utils
