+++ packages/core/src/address-utils/index.ts
@@ -1,0 +1,1 @@
+export * from './address-utils';
